docling_rag/pyproject.toml
[project]
name = "docling-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "args>=0.1.0",
    "docling>=2.28.0",
    "docling-core>=2.23.3",
    "langchain-huggingface>=0.1.2",
    "langchain-milvus>=0.1.8",
    "langchain-ollama>=0.3.0",
    "langchain-text-splitters>=0.3.7",
    "python-dotenv>=1.0.1",
]
